feat: refactor Feed code
