--- src/routes/Feed/FeedHeader/index.ts
@@ -1,1 +1,0 @@
-export { default as FeedHeader } from "./FeedHeader"
