+++ src/routes/Feed/LeftSidebar/index.tsx
@@ -1,0 +1,13 @@
+import { FC } from "react"
+import TagList from "../components/TagList"
+import { wrapperStyle } from "./styles"
+
+const LeftSidebar: FC = () => {
+  return (
+    <div css={wrapperStyle}>
+      <TagList />
+    </div>
+  )
+}
+
+export default LeftSidebar
