+++ src/routes/Feed/components/Footer.tsx
@@ -1,6 +1,6 @@
-import { CONFIG } from "site.config"
-import React from "react"
 import styled from "@emotion/styled"
+import React from "react"
+import { CONFIG } from "site.config"
 
 const d = new Date()
 const y = d.getFullYear()
